Implement a Hiera 5 environment hiera.yaml file
This hiera.yaml configures where the environment expects data to be
found. The vRO starter content repo in particular is dependent on
common.yaml, so it's a good idea to make sure automatic parameter
lookups are actually consulting that file.

In previous versions of Puppet the default out-of-box global hierarchy
happened to injest this file. With the advent of Hiera 5 though, that
changes, and it is expected that a Puppet environment will
self-reference data it needs or expects via its own hiera.yaml (the
global hiera.yaml is now just an override).